cs4-rag-search/requirements.txt
litellm>=1.30.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
httpx>=0.27.0
structlog>=24.1.0
sse-starlette>=2.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
